+++ game/scripts/vscripts/events.lua
@@ -1,7 +1,25 @@
 require('modifiers/player/enable_player_modifier')
-require('event/js_event')
-require('event/creep')
 require('event/kill')
+
+function AIGameMode:OnGetLoadingSetOptions(eventSourceIndex, args)
+    if tonumber(args.host_privilege) ~= 1 then
+        return
+    end
+    self.iDesiredRadiant = tonumber(args.game_options.radiant_player_number)
+    self.iDesiredDire = tonumber(args.game_options.dire_player_number)
+    self.fPlayerGoldXpMultiplier = tonumber(args.game_options.player_gold_xp_multiplier)
+    self.fBotGoldXpMultiplier = tonumber(args.game_options.bot_gold_xp_multiplier)
+
+    self.iRespawnTimePercentage = tonumber(args.game_options.respawn_time_percentage)
+    self.iMaxLevel = tonumber(args.game_options.max_level)
+
+    self.iTowerPower = tonumber(args.game_options.tower_power)
+
+    self.iStartingGoldPlayer = tonumber(args.game_options.starting_gold_player)
+    self.iStartingGoldBot = tonumber(args.game_options.starting_gold_bot)
+    self.bSameHeroSelection = args.game_options.same_hero_selection
+    self:PreGameOptions()
+end
 
 function AIGameMode:InitPlayerGold()
     if self.PreGameOptionsSet then
@@ -299,6 +317,20 @@ function AIGameMode:OnNPCSpawned(keys)
 
         hEntity.bInitialized = true
     end
+end
+
+-- 小兵金钱随时间增加
+function AIGameMode:GetLaneGoldMul()
+    local time = GameRules:GetDOTATime(false, false)
+    local mul = 1
+    if time <= 15 * 60 then
+        mul = 1
+    else
+        mul = math.floor(time / 900)
+    end
+    -- 60分钟封顶
+    mul = math.min(mul, 4)
+    return mul
 end
 
 function AIGameMode:OnItemPickedUp(event)
